requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Pillow==10.1.0
python-slugify==8.0.1
bleach==6.1.0


